Added three new classes to the project: 	NarrowSelection 	WiddenSelection 	XmlHelpers
The first two are just shells at this time, but will be developed into commands to reduce and to encrease the size of a selection. The third contains extensions for testing Xml selections to ensure that the selected text selects complete controls. These extensions will be used in the other classes, but have already been employed to improve validity checking in several existing commands.